+++ src/euroexa_testbed2/libxdma.c
@@ -355,7 +355,7 @@ xdma_status xdmaFree(xdma_buf_handle handle) {
     return ret;
 }
 
-static inline xdma_status _xdmaStream(xdma_buf_handle buffer, size_t len, unsigned int offset,
+static inline xdma_status _xdmaStream(xdma_buf_handle buffer, size_t len, size_t offset,
         xdma_device dev, xdma_channel ch, bool block, xdma_transfer_handle *transfer)
 {
     struct xdma_chan_cfg *channel = (struct xdma_chan_cfg*)ch;
@@ -409,20 +409,20 @@ static inline xdma_status _xdmaStream(xdma_buf_handle buffer, size_t len, unsign
     return XDMA_SUCCESS;
 }
 
-xdma_status xdmaStream(xdma_buf_handle buffer, size_t len, unsigned int offset,
+xdma_status xdmaStream(xdma_buf_handle buffer, size_t len, size_t offset,
             xdma_device dev, xdma_channel channel)
 {
     return _xdmaStream(buffer, len, offset, dev, channel, 1 /*block*/, NULL /*xdma_transfer_handle*/);
 }
 
-xdma_status xdmaStreamAsync(xdma_buf_handle buffer, size_t len, unsigned int offset,
+xdma_status xdmaStreamAsync(xdma_buf_handle buffer, size_t len, size_t offset,
         xdma_device dev, xdma_channel channel, xdma_transfer_handle *transfer)
 {
     return _xdmaStream(buffer, len, offset, dev, channel, 0 /*block*/, transfer);
 }
 
 #define USE_CDMA
-xdma_status xdmaMemcpy(void *usr, xdma_buf_handle buffer, size_t len, unsigned int offset,
+xdma_status xdmaMemcpy(void *usr, xdma_buf_handle buffer, size_t len, size_t offset,
         xdma_dir mode)
 {
 #ifdef USE_CDMA
@@ -444,7 +444,7 @@ xdma_status xdmaMemcpy(void *usr, xdma_buf_handle buffer, size_t len, unsigned i
 
 
 xdma_status xdmaMemcpyAsyncChunk(void *usr, xdma_buf_handle buffer, size_t len,
-        unsigned int offset, xdma_dir mode, xdma_transfer_handle *transfer)
+        size_t offset, xdma_dir mode, xdma_transfer_handle *transfer)
 {
 //    *transfer = (xdma_transfer_handle)NULL;
 //    return xdmaMemcpy(usr, buffer, len, offset, mode);
@@ -496,7 +496,7 @@ xdma_status xdmaMemcpyAsyncChunk(void *usr, xdma_buf_handle buffer, size_t len,
 }
 
 xdma_status xdmaMemcpyAsync(void *usr, xdma_buf_handle buffer, size_t len,
-        unsigned int offset, xdma_dir mode, xdma_transfer_handle *transfer) {
+        size_t offset, xdma_dir mode, xdma_transfer_handle *transfer) {
     int rem = len;
     int transferred = 0;
     xdma_transfer_handle tmpHandle = 0;
